+++ library/components/taskdetail/blocks/sections/Attachment/Attachment.vue
@@ -1,5 +1,5 @@
 <script setup lang="ts">
-import { inject, onMounted, Ref, ref } from 'vue';
+import { computed, ComputedRef, inject, onMounted, Ref, ref } from 'vue';
 import Icon from 'lib/components/icon/Icon.vue';
 import Button from 'lib/components/button/Button.vue';
 import AttachmentItem from './AttachmentItem.vue';
@@ -8,6 +8,8 @@ import DialogAddAttachment from '../../common/DialogAddAttachment.vue';
 import TaskAttachmentServices from 'lib/services/taskAttachment.service';
 import AttachmentChangelog from './AttachmentChangelog.vue';
 import { useToast } from 'lib/utils';
+import { WangsitStatus } from 'lib/types/wangsStatus.type';
+import { TaskDetailData } from 'lib/types/task.type';
 
 const toast = useToast();
 
@@ -15,12 +17,25 @@ onMounted(() => {
   getAttachments();
 });
 
+const userType =
+  inject<ComputedRef<'member' | 'admin' | 'pm' | 'teamLeader' | 'guest'>>(
+    'userType',
+  );
+const taskDetail = inject<Ref<TaskDetailData>>('taskDetail');
 const taskId = inject<Ref<string>>('taskId');
 
 const attachments = ref<AttachmentItemData[]>([]);
 
 const dialogAddAttachment = ref(false);
 const dialogChangelog = ref(false);
+
+const isDisabled = computed(() => {
+  const disabledStatus = (
+    ['Selesai', 'Reported Bug'] as WangsitStatus[]
+  ).includes(taskDetail.value?.status);
+
+  return disabledStatus || userType.value === 'guest';
+});
 
 const getAttachments = async (): Promise<void> => {
   try {
@@ -54,6 +69,7 @@ const getAttachments = async (): Promise<void> => {
           text
         />
         <Button
+          :disabled="isDisabled"
           @click="dialogAddAttachment = true"
           icon="add"
           label="Attachment"